+++ include/AGEUN.h
@@ -6,5 +6,5 @@
 class AGEUN : public AGE
 {
 public:
-	std::vector<int> execute();
+	std::vector<float> execute();
 };
