+++ include/AGGUN.h
@@ -6,5 +6,5 @@
 class AGGUN : public AGG
 {
 public:
-	std::vector<int> execute();
+	std::vector<float> execute();
 };
