+++ src/AGESF.cpp
@@ -1,3 +1,3 @@
 #include "AGESF.h"
 
-std::vector<int> AGESF::execute() { return executeGeneticAlgoritm(2, 1); }
+std::vector<float> AGESF::execute() { return executeGeneticAlgoritm(2, 1); }
