+++ src/AGEUN.cpp
@@ -1,3 +1,3 @@
 #include "AGEUN.h"
 
-std::vector<int> AGEUN::execute() { return executeGeneticAlgoritm(2, 0); }
+std::vector<float> AGEUN::execute() { return executeGeneticAlgoritm(2, 0); }
